Provide variables "gui::horPad" and "gui::vertPad" to GUI scripts
they are set before onActivate and also whenever the (Doom3 SDL) window
size changes, in that case additionally the "UpdateWindowSize" Named
Event is called in the script, in case you wanna do anything special
in that case:

  onNamedEvent UpdateWindowSize {
    set "print" "UpdateWindowSize - horPad:" "gui::horPad"
          "vertPad:" "gui::vertPad";
  }

horPad is the width of the padding border (empty space) between the
  left/right border of Doom3's SDL window and a centered
  (cstAnchor CST_ANCHOR_CENTER_CENTER) full-sized windowDef in virtual
  pixels (in the Doom3 GUI coordinate system, which usually is 640x480).
vertPad is the same for the padding border between the top/bottom border
  and a centered full-sized windowDef

One of those values will always be 0.

So for example if you play Doom3 at 1600x1200 (a 4:3 resolution like
640x480) both horPad and vertPad are 0.
If you play at 1680x1050 (16:10 resolution) and your windowDef Desktop
is 640x480 (the default), horPad is 64 (and vertPad is  0), because
there are 64 virtual pixels between the left window border and a
centered 640x480 windowDef (and also between the centered windowDef
and the right window border.

gui::horPad and gui::vertPad are useful for filling those otherwise
empty padding areas with windowDefs that you anchor to the corresponding
points of the window (like CST_ANCHOR_LEFT and CST_ANCHOR_RIGHT when
there's horizontal padding) that have just the right size.

Example of a green rectangle filling parts of the left padding area,
but not touching the center area (10 virtual pixels distance):

  windowDef leftPadFiller {
    cstAnchor  CST_ANCHOR_LEFT
    rect   0, 200, "gui::horPad" - 10, 40
    backColor 0, 1, 0, 1
  }